feat/#163 컴포넌트 분리 및 layout으로 감쌈
